refactor(ucmc-web): extract pending + lifecycle tab views into components
`PendingTab` (date-range filter + bulk approve/reject) and `LifecycleTab`
(shared between rejected and deactivated, parameterized by `kind`) move
out of `routes/members.management.tsx` into `features/members/components/`,
matching the existing `UnclaimedTab` shape. URL state stays owned by the
route and threads through props. No user-visible change.
